+++ util/CHANGELOG.md
@@ -5,7 +5,9 @@ Refer to https://keepachangelog.com/en/1.0.0/ for how to maintain changelog<br /
 
 ## [Unreleased]
 
--
+-   `copyDirRecursive` debug logs don't show full paths, only basename
+-   added `isProd` util
+-   debug logs won't show without LOG_DEBUG var & will never show in production
 
 ## [0.3.0] - 03/05/2023